requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.47.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
pypdf==4.3.1
python-docx==1.1.2
reportlab==4.2.5
